перетащил из хендлеров код напоминалки reminder_handler теперь чисто todo скрипт))
