+++ handlers/laundry_handler.py
@@ -25,7 +25,11 @@ from services.laundry.plot_schedule import plot_schedule
 from services.laundry.schedule import Schedule
 from services.wallet.wallet import get_balance, debit_balance, credit_balance
 
-from handlers.reminder_handler import add_reminders
+import sys
+from pathlib import Path
+sys.path.append(str(Path(__file__).parent.parent))
+
+from reminder.reminder_tools import add_reminders
 
 laundry_router = Router()
